Adds speed counter and its corresponding unit test
The speed counter simply requires that a triggering function is called
to indicate pulses of an event. The rate of the pulses can be returned
in units of per-second (Hertz), per-minute, and per-hour.

The corresponding unit test makes subtle changes to its corresponding
behavior based on the platform the test is running on. In this commit,
the unit tests have only been tested on a Windows platform. A unix
machine hypothetically keeps tighter tolerances on its timing,
especially on a system with so little going on such as a raspberry pi,
so this branch may need to be revisted once the code has been unit
tested on the raspberry pi.