No longer OS dependent, added more assertions
With this commit, a new method of testing the behavior of a trigger has
been implemented that removes the dependency on the operating system's
timing precision. Rather than using a busy loop or a time.sleep, an
initial time is simply injected, and then the speed counter is triggered
to check the frequency counting.

Some other assertions are checked now as well to ensure the time values
in the speed counter have been appropriately set.